+++ src/libtrx/include/libtrx/game/sound/common.h
@@ -8,9 +8,6 @@
 
 #include <stddef.h>
 
-#define SOUND_MIN_VOLUME 0
-#define SOUND_MAX_VOLUME 10
-
 // Stops and unloads all samples
 extern void Sound_Reset(void);
 
@@ -24,7 +21,7 @@ SAMPLE_INFO *Sound_GetSampleInfo(SOUND_EFFECT_ID sfx_num);
 SAMPLE_INFO *Sound_GetSampleInfoByIdx(int32_t info_idx);
 
 extern int32_t Sound_GetMasterVolume(void);
-extern void Sound_SetMasterVolume(int32_t volume);
+extern void Sound_SetMasterVolume(float volume);
 
 void Sound_ResetSources(void);
 void Sound_PauseAll(void);