+++ src/tr2/game/fmv.c
@@ -117,10 +117,7 @@ static bool M_Play(const char *const file_name)
 
     Video_Start(video);
     while (video->is_playing) {
-        Video_SetVolume(
-            video,
-            m_Muted ? 0
-                    : g_Config.audio.sound_volume / (float)SOUND_MAX_VOLUME);
+        Video_SetVolume(video, m_Muted ? 0.0f : g_Config.audio.sound_volume);
 
         const SHELL_SIZE display_size = Shell_GetCurrentDisplaySize();
         Video_SetSurfaceSize(video, display_size.w, display_size.h);